Loading page in all sites, and eliminate 14 class from routes and pages, zip in 2 class
