+++ frontend/.tanstack/tmp/c8e0c333-0e97529407cab7ae8833bd45fec99523
@@ -1,8 +1,8 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { useQuery } from "@tanstack/react-query";
-import SculturePage from "../components/pages/SculturePage";
+import AnimationPage from "../components/pages/AnimationPage";
 
-export const Route = createFileRoute("/Sculture")({
+export const Route = createFileRoute("/Animation")({
   component: RouteComponent,
 });
 
@@ -28,7 +28,15 @@ function RouteComponent() {
   if (!token)
     return <p className="text-white text-center mt-10">No estás logueado.</p>;
   if (isLoading)
-    return <p className="text-white text-center mt-10">Loading...</p>;
+    return (
+      <div className="flex bg-stone-950 text-white items-center h-full justify-center pb-20">
+        <div className="flex space-x-3">
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce [animation-delay:-0.6s]"></div>
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce [animation-delay:-0.3s]"></div>
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce "></div>
+        </div>
+      </div>
+    );
   if (error)
     return (
       <p className="text-red-500 text-center mt-10">
@@ -42,10 +50,10 @@ function RouteComponent() {
       new Date(b.created_at).getTime() - new Date(a.created_at).getTime()
   );
 
-  const sculturePublications = sortedPublications.filter(
+  const animationPublications = sortedPublications.filter(
     (pub: { category?: string }) =>
-      pub.category?.toLowerCase() === "Sculture".toLowerCase()
+      pub.category?.toLowerCase() === "Animation".toLowerCase()
   );
 
-  return <SculturePage sculturePublications={sculturePublications} />;
+  return <AnimationPage animationPublications={animationPublications} />;
 }
