+++ frontend/.tanstack/tmp/c8e0c333-32bb155755187e6895c02d6f24afb9cf
@@ -1,0 +1,10 @@
+import { createFileRoute } from "@tanstack/react-router";
+import Home from "../components/pages/IndexPage";
+
+export const Route = createFileRoute("/")({
+  component: HomeRoute,
+});
+
+function HomeRoute() {
+  return <Home subtitle="“Tu arte, tu espacio protegido”" title="HARIXOM" />;
+}
