+++ frontend/.tanstack/tmp/c8e0c333-9b97e2c4ed09e1830a32e884d1de730c
@@ -1,16 +1,17 @@
-import { createFileRoute } from "@tanstack/react-router";
+import { createFileRoute, useParams } from "@tanstack/react-router";
 import { useQuery } from "@tanstack/react-query";
-import Art3DPage from "../components/pages/3DPage";
+import Categories from "../components/pages/Categories";
 
-export const Route = createFileRoute("/3d")({
+export const Route = createFileRoute("/Categories")({
   component: RouteComponent,
 });
 
 function RouteComponent() {
   const token = localStorage.getItem("access_token");
+  const { name } = useParams({ from: "/Categories/$name" });
 
   const { data, isLoading, error } = useQuery({
-    queryKey: ["publications", "3d"],
+    queryKey: ["allPublications"],
     queryFn: async () => {
       const response = await fetch("http://127.0.0.1:8000/api/publications", {
         headers: {
@@ -28,7 +29,15 @@ function RouteComponent() {
   if (!token)
     return <p className="text-white text-center mt-10">No estás logueado.</p>;
   if (isLoading)
-    return <p className="text-white text-center mt-10">Loading...</p>;
+    return (
+      <div className="flex bg-stone-950 text-white items-center h-full justify-center pb-20">
+        <div className="flex space-x-3">
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce [animation-delay:-0.6s]"></div>
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce [animation-delay:-0.3s]"></div>
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce "></div>
+        </div>
+      </div>
+    );
   if (error)
     return (
       <p className="text-red-500 text-center mt-10">
@@ -42,10 +51,16 @@ function RouteComponent() {
       new Date(b.created_at).getTime() - new Date(a.created_at).getTime()
   );
 
-  const Art3DPublications = sortedPublications.filter(
+  const categoryPublications = sortedPublications.filter(
     (pub: { category?: string }) =>
-      pub.category?.toLowerCase() === "3D Art".toLowerCase()
+      pub.category?.toLowerCase() === "Photography".toLowerCase()
   );
 
-  return <Art3DPage art3DPublications={Art3DPublications} />;
+  return <Categories
+    title={name}
+    style="text-[#ff6161]"
+    icon="./public/icon-foto.svg"
+    altIcon="photography icon"
+    categoriesPublications={categoryPublications}
+  />;
 }
