+++ frontend/src/routes/Profile.tsx
@@ -50,7 +50,15 @@ function ProfileRoute() {
   if (!token)
     return <p className="text-white text-center mt-10">No estás logueado.</p>;
   if (isLoading)
-    return <p className="text-white text-center mt-10">Loading...</p>;
+    return (
+      <div className="flex bg-stone-950 text-white items-center h-full justify-center pb-20">
+        <div className="flex space-x-3">
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce [animation-delay:-0.6s]"></div>
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce [animation-delay:-0.3s]"></div>
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce "></div>
+        </div>
+      </div>
+    );
   if (error)
     return (
       <p className="text-red-500 text-center mt-10">
