+++ frontend/src/routes/ProfileGuest.tsx
@@ -50,7 +50,15 @@ function ProfileGuestRoute() {
   
 
   if (!guestUserId) return <p className="text-white text-center mt-10">Usuario no especificado.</p>;
-  if (isLoading) return <p className="text-white text-center mt-10">Loading...</p>;
+  if (isLoading) return (
+      <div className="flex bg-stone-950 text-white items-center h-full justify-center pb-20">
+        <div className="flex space-x-3">
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce [animation-delay:-0.6s]"></div>
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce [animation-delay:-0.3s]"></div>
+          <div className="w-4 h-4 bg-blue-500 rounded-full animate-bounce "></div>
+        </div>
+      </div>
+    );
   if (error) return <p className="text-red-500 text-center mt-10">{(error as Error).message}</p>;
 
   const user = profileData.user;
